+++ src/dockerfiles/Automattic__wp-calypso-23915.Dockerfile
@@ -91,7 +91,7 @@ python2 -V
 EOF_5b48db3d125f
 
 
-RUN <<EOF_3ac153fb6ab7
+RUN <<EOF_094b1b586dcf
 #!/bin/bash
 set -euxo pipefail
 (mkdir -p /testbed && cd /testbed && git init -q . && git remote add origin https://github.com/Automattic/wp-calypso && git fetch -q --depth 1 origin 24d762a0fe3a92500f92323b038a2389cb6ecae5 && git reset -q --hard FETCH_HEAD) || (rm -rf /testbed && git clone -o origin https://github.com/Automattic/wp-calypso /testbed)
@@ -99,16 +99,20 @@ chmod -R 777 /testbed
 cd /testbed
 git reset --hard 24d762a0fe3a92500f92323b038a2389cb6ecae5
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 24d762a0fe3a92500f92323b038a2389cb6ecae5)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+TARGET_TIMESTAMP=$(git show -s --format=%ci 24d762a0fe3a92500f92323b038a2389cb6ecae5)
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+AFTER_TIMESTAMP=$(date -d "$TARGET_TIMESTAMP + 1 second" '+%Y-%m-%d %H:%M:%S')
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install --unsafe-perm
-EOF_3ac153fb6ab7
+EOF_094b1b586dcf
 
 
 RUN <<EOF_abda33051799
