+++ src/dockerfiles/Automattic__wp-calypso-33948.Dockerfile
@@ -91,7 +91,7 @@ python2 -V
 EOF_16a722167964
 
 
-RUN <<EOF_34888c0669ca
+RUN <<EOF_418caa49f752
 #!/bin/bash
 set -euxo pipefail
 (mkdir -p /testbed && cd /testbed && git init -q . && git remote add origin https://github.com/Automattic/wp-calypso && git fetch -q --depth 1 origin d1c128c3c97e81421f8137c138a2a1b112d44c8b && git reset -q --hard FETCH_HEAD) || (rm -rf /testbed && git clone -o origin https://github.com/Automattic/wp-calypso /testbed)
@@ -99,16 +99,20 @@ chmod -R 777 /testbed
 cd /testbed
 git reset --hard d1c128c3c97e81421f8137c138a2a1b112d44c8b
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct d1c128c3c97e81421f8137c138a2a1b112d44c8b)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+TARGET_TIMESTAMP=$(git show -s --format=%ci d1c128c3c97e81421f8137c138a2a1b112d44c8b)
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+AFTER_TIMESTAMP=$(date -d "$TARGET_TIMESTAMP + 1 second" '+%Y-%m-%d %H:%M:%S')
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm install --unsafe-perm
-EOF_34888c0669ca
+EOF_418caa49f752
 
 
 RUN <<EOF_0ff77a395366
